revert back to 0dd82f1, to remove brokers in parse.
If any local variable in a block is referred as upvalue, it should
be closed when the block ends. This is what CloseBlock does.
However if the block is a loop (e.g. `for` or `while`), the local
variable should be referred as *independent* upvalue multiple times.
It's dynamically! It can only be created when executing in VM, but
*not* in parsing.