feat: add missing usize return -> 2-1
the function originally is set as follow:
```rust
fn add_const(constants: &mut Vec<Value>, c: Value) -> usize {
     constants. push(c)
}
```
after that, the reader is conditioned to run the program in order to test the program working and see the outputs, but a return is missing;
What means the code doesn't compile. This fixes it by returning constants length as it should.

a better implementation is made later on chapter, but for a better reader experience, this commit should be valid.